Revert "SplitView: restore the server buffer"
This reverts commit f906793d72c51b38fefb5b185eb1b971aa89421f.

The server buffer is restored based on its title, there's no need for
a separate server name check.

Also, the code was also causing every channel up until the one that
matched the title to be made the current buffer of the view, as the
server name of the channel and the server name of the buffer view would
always match, but we only reset the '__unrestored__' object name on a
title match. The result was that the channels were made visible, and
hence would result in unread counts being reset.